Update style.css
To be honest, I don't really know if I did it right, but I tried to use the 'pearl river' color as background color (supposed to be a javascript color). I hope it works, but I have no experience or knowledge in coding so don't expect anything, I don't know what I was doing.